added kafka producer rights

+++ common-service/kafka-producer/src/main/java/org/go/together/kafka/producer/config/ProducerKafkaConfig.java
@@ -1,6 +1,7 @@
 package org.go.together.kafka.producer.config;
 
 import org.go.together.dto.Dto;
+import org.go.together.kafka.producer.beanpostprocessor.ProducerRights;
 import org.go.together.kafka.producer.config.crud.CreateProducerKafkaConfig;
 import org.springframework.beans.factory.config.ConfigurableListableBeanFactory;
 
@@ -16,12 +17,25 @@ public abstract class ProducerKafkaConfig<D extends Dto> extends CreateProducerK
 
     public void configureProducer(String kafkaServer,
                                   String kafkaGroupId,
-                                  ConfigurableListableBeanFactory beanFactory) {
-        createProducerBeanFactoryPostProcessor(kafkaServer, kafkaGroupId, beanFactory);
-        deleteBeanFactoryPostProcessor(kafkaServer, beanFactory);
-        findProducerBeanFactoryPostProcessor(kafkaServer, kafkaGroupId, beanFactory);
-        readProducerBeanFactoryPostProcessor(kafkaServer, kafkaGroupId, beanFactory);
-        updateProducerBeanFactoryPostProcessor(kafkaServer, kafkaGroupId, beanFactory);
-        validateProducerBeanFactoryPostProcessor(kafkaServer, kafkaGroupId, beanFactory);
+                                  ConfigurableListableBeanFactory beanFactory,
+                                  ProducerRights producerRights) {
+        if (producerRights.isCreate()) {
+            createProducerBeanFactoryPostProcessor(kafkaServer, kafkaGroupId, beanFactory);
+        }
+        if (producerRights.isDelete()) {
+            deleteBeanFactoryPostProcessor(kafkaServer, beanFactory);
+        }
+        if (producerRights.isFind()) {
+            findProducerBeanFactoryPostProcessor(kafkaServer, kafkaGroupId, beanFactory);
+        }
+        if (producerRights.isRead()) {
+            readProducerBeanFactoryPostProcessor(kafkaServer, kafkaGroupId, beanFactory);
+        }
+        if (producerRights.isUpdate()) {
+            updateProducerBeanFactoryPostProcessor(kafkaServer, kafkaGroupId, beanFactory);
+        }
+        if (producerRights.isValidate()) {
+            validateProducerBeanFactoryPostProcessor(kafkaServer, kafkaGroupId, beanFactory);
+        }
     }
 }
